Move GF filters from functions.php to GravityFormsExtensions.php
This is to keep all Gravity Forms code in one place
