+++ functions.php
@@ -214,77 +214,6 @@ add_filter(
 );
 
 /**
- *
- * Add CSS classes to Gravity Forms fields.
- */
-add_filter( 'gform_field_css_class', 'gform_custom_field_class', 10, 3 );
-
-/**
- *
- * Add CSS classes to some Gravity Forms fields: checkboxes and radio buttons.
- *
- * @param string $classes The existing field classes.
- * @param object $field The field.
- *
- * @return string The updated field classes.
- */
-function gform_custom_field_class( $classes, $field ) {
-	if ( 'checkbox' === $field->type || 'radio' === $field->type || 'consent' === $field->type ) {
-		$classes .= ' custom-control';
-	}
-	return $classes;
-}
-
-/**
- *
- * Update confirmation message for Gravity Forms.
- */
-add_filter( 'gform_confirmation', 'gform_custom_confirmation', 10, 3 );
-
-/**
- *
- * Return custom confirmation message for Gravity Forms.
- *
- * @param string $confirmation The default confirmation message.
- * @param array  $form The form properties.
- *
- * @return string The custom confirmation message.
- */
-function gform_custom_confirmation( $confirmation, $form ) {
-	// If the $confirmation object is an array, it means that it's a redirect page so we can directly use it.
-	if ( is_array( $confirmation ) ) {
-		return $confirmation;
-	}
-
-	$post = Timber::query_post( false, Post::class );
-
-	$default_confirmation_array = array_filter(
-		$form['confirmations'],
-		function ( $conf ) use ( $confirmation ) {
-			return str_contains( $confirmation, $conf['message'] );
-		}
-	);
-
-	$default_confirmation = reset( $default_confirmation_array );
-
-	$confirmation_fields = [
-		'confirmation'    => $default_confirmation['message'],
-		'share_platforms' => [
-			'facebook' => '1' === $default_confirmation['facebook'] ?? false,
-			'twitter'  => '1' === $default_confirmation['twitter'] ?? false,
-			'whatsapp' => '1' === $default_confirmation['whatsapp'] ?? false,
-			'native'   => '1' === $default_confirmation['native'] ?? false,
-			'email'    => '1' === $default_confirmation['email'] ?? false,
-		],
-		'share_text'      => $default_confirmation['p4_gf_share_text_override'] ?? '',
-		'share_url'       => $default_confirmation['p4_gf_share_url_override'] ?? '',
-		'post'            => $post,
-	];
-
-	return Timber::compile( [ 'gravity_forms_confirmation.twig' ], $confirmation_fields );
-}
-
-/**
  * TODO: Move to editor only area.
  * Set the editor width per post type.
  */
